chore: Update pr comment behavior (#47)
Currently every pr commit results in a new comment like
https://github.com/neuralmagic/speculators/pull/39#issuecomment-3020058659
which clutters the pr.

This pr changes the behavior so that new builds edit the existing
comment instead of creating a new one. Previous comments can be seen by
looking at the revision history of the comment.

Note: this uses the existing Github token, instead of the automation bot
token but could be updated to use that token instead if that's
preferred.

Signed-off-by: Fynn Schmitt-Ulms <fschmitt@redhat.com>