Remove remaining python 3.9 usages (#152)
I tried to just update the 3.9 usages and it resulted in a bunch of new
ruff lint errors.

Setting the target version for `ruff` to `3.10` resulted in a bunch of
upgrade lints failing. Most of these were automatically handled by `ruff
check --fix` and `ruff format`. A few required manual intervention.

Lastly, I updated the version of ruff we use because there was an
upgrade rule on the old version that has since been deprecated. I
could've added an ignore rule instead but I don't think it hurts here to
update to a newer ruff version.

---------

Signed-off-by: Fynn Schmitt-Ulms <fschmitt@redhat.com>