Only load Verifier model if attachment_mode is 'full' (#154)
Currently, if `verifier_attachment_mode` is `"full"` or `"train_only"`
we load the full verifier model on init.

However, for the `"train_only"` case, we want to only load the parts of
the model that are needed. The logic to do so, is meant to reside in the
subclasses of `SpeculatorModel`.

This change, makes it so that we don't auto-load the full verifier model
if `"train_only"` is selected.
Additional changes:
- `SpeculatorModel.attach_verifier()` no longer returns the loaded
verifier (since it might not exist)
- `EagleSpeculator.attach_verifier()` directly calls
`resolve_verifier()` to load the verifier model and extract its
components.
- `test_speculator_model_attach_verifier_invalid` needed to be updated
to reset the model between subtests because otherwise the
`verifier_attachment_mode` gets set and kept by the previous try-excepts
blocks.

Signed-off-by: Fynn Schmitt-Ulms <fschmitt@redhat.com>