Separate definitions of first decoder block layer from subsequent
In the Eagle3 Algorithm, the first layer needs to be modified to support a larger (2x) hidden dim, while subsequent layers behave as regular.
Previously, we used a special decoder layer class that behaved differently depending on the `layer_idx` it received. Now we instead, use the 
special class only for the first layer and switch to using the original `LlamaDecoderLayer` for subsequent layers. 


Signed-off-by: Fynn Schmitt-Ulms <fschmitt@redhat.com>