Fix link redirects (#205)
Closes #192 

There are a lot of links in the repo that reference the old repo
location https://github.com/neuralmagic/speculators as well as other
links to third-party websites that go through several redirects.

This pr replaces those links with links directly to the final
destination.


Note the only redirect left after this pr is pypi's:
Redirects: https://docs.pytest.org/ -->
https://docs.pytest.org/en/stable/

This is intentional, so that users with different local languages will
be redirect to their local language option.

Signed-off-by: Fynn Schmitt-Ulms <fschmitt@redhat.com>